+++ src/dyscord/__init__.py
@@ -1,6 +1,6 @@
 '''Dyscord is a library to enable a reasonably 1:1 binding between python and the official API.'''
 
-from . import client, helper, objects, utilities
+from . import client, helper, objects, utilities, command
 from .version import __version__
 from .client.discord_client import DiscordClient
 
@@ -11,4 +11,5 @@ __all__ = [
     'helper',
     'objects',
     'utilities',
+    'command',
 ]